Bug Fix: Create .reg file in $PSScriptRoot dir.
Discovered when debugging Create-Randomize-AccentColorJob script.
"none.reg" was previously being created in the directory from which the
script was invoked, not $PSScriptRoot. This was problematic when we
invoked Randomize-MailAppAccentColor as a part of "powershell -command" as
we tried to promptly create none.reg within the PowerShell source
directory. This caused permissions issues.

Note that Create-RandomizeAccentColorJob still doesn't work though; there
are still other issues to be debugged.